refactor: move all database concepts into a concepts folder.
