+++ packages/serverpod/lib/database.dart
@@ -1,11 +1,11 @@
 library database;
 
-export 'src/database/columns.dart';
+export 'src/database/concepts/columns.dart';
 export 'src/database/database_pool_manager.dart';
-export 'src/database/expressions.dart';
-export 'src/database/includes.dart';
-export 'src/database/many_relation.dart';
-export 'src/database/order.dart';
-export 'src/database/table.dart';
-export 'src/database/transaction.dart';
+export 'src/database/concepts/expressions.dart';
+export 'src/database/concepts/includes.dart';
+export 'src/database/concepts/many_relation.dart';
+export 'src/database/concepts/order.dart';
+export 'src/database/concepts/table.dart';
+export 'src/database/concepts/transaction.dart';
 export 'src/generated/database/enum_serialization.dart';
