+++ packages/serverpod/lib/src/database/concepts/table.dart
@@ -1,5 +1,5 @@
 import 'package:serverpod/database.dart';
-import 'package:serverpod/src/database/table_relation.dart';
+import 'package:serverpod/src/database/concepts/table_relation.dart';
 import 'package:serverpod_serialization/serverpod_serialization.dart';
 
 /// Holds data corresponding to a row in the database. Concrete classes are
