+++ packages/serverpod/lib/src/database/concepts/table_relation.dart
@@ -1,7 +1,7 @@
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
-import 'package:serverpod/src/database/columns.dart';
-import 'package:serverpod/src/database/table.dart';
+import 'package:serverpod/src/database/concepts/columns.dart';
+import 'package:serverpod/src/database/concepts/table.dart';
 import 'package:serverpod_shared/serverpod_shared.dart';
 
 /// Called for all query aliases generated by [TableRelation].
