+++ packages/serverpod/test/database/table_relation_test.dart
@@ -1,6 +1,6 @@
-import 'package:serverpod/src/database/columns.dart';
-import 'package:serverpod/src/database/table.dart';
-import 'package:serverpod/src/database/table_relation.dart';
+import 'package:serverpod/src/database/concepts/columns.dart';
+import 'package:serverpod/src/database/concepts/table.dart';
+import 'package:serverpod/src/database/concepts/table_relation.dart';
 import 'package:serverpod/test_util/table_relation_builder.dart';
 import 'package:serverpod_shared/serverpod_shared.dart';
 import 'package:test/test.dart';
